Add files via upload
CHEP

CHEP (Clustering of HElical Proteins) is a python  program for the clustering of 2D helical proteins into homogeneous groups by using the results of 2D classification and information about the association of segments to the helical proteins. It is developed in the group of Gunnar Schröder at the Forschungszentrum Jülich.

The k-means clustering and elbow tests are implemented using Sklearn modules. The results of this algorithm will be published soon.